requirements.txt
streamlit>=1.45.0
openai>=1.40.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.1
plotly>=5.24.0
pydantic>=2.8.0
scikit-learn>=1.5.0
yfinance>=0.2.54
